refactor: 重构读公告服务的实现
